# dashboard/src/api/models/uIBinding.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { UIBindingMode } from './uIBindingMode';
import type { BindingScope } from './bindingScope';

export interface UIBinding {
  /** e.g. parameters.url / results.status */
  path: string;
  mode: UIBindingMode;
  /** Optional textual prefix (e.g. '@welcomeJourney.stage2.#notifyCustomer') that scopes the binding before the JSON pointer root. */
  prefix?: string;
  scope?: BindingScope;
}