# dashboard/src/api/models/workflowMiddleware.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { WorkflowNodeRole } from './workflowNodeRole';
import type { NodePackage } from './nodePackage';
import type { WorkflowNodeStatus } from './workflowNodeStatus';
import type { WorkflowNodeParameters } from './workflowNodeParameters';
import type { WorkflowNodeResults } from './workflowNodeResults';
import type { WorkflowNodeState } from './workflowNodeState';
import type { WorkflowNodeSchema } from './workflowNodeSchema';
import type { NodeUI } from './nodeUI';

export interface WorkflowMiddleware {
  /** Middleware UUID */
  id: string;
  /** e.g. "middleware.loop" */
  type: string;
  /** Execution role of the middleware (always middleware). */
  role?: WorkflowNodeRole;
  package: NodePackage;
  /** Middleware lifecycle state. */
  status: WorkflowNodeStatus;
  /** Group/category shown in the builder palette. */
  category: string;
  label: string;
  /** Longer description of the middleware behaviour. */
  description?: string;
  /** Keywords for search/filter. */
  tags?: string[];
  /** Default parameter payload seeded from the manifest schema. */
  parameters?: WorkflowNodeParameters;
  /** Default results payload seeded from the manifest schema. */
  results?: WorkflowNodeResults;
  /** Platform-managed execution hints surfaced in the builder (optional). */
  state?: WorkflowNodeState;
  /** JSON Schemas for parameters/results. */
  schema?: WorkflowNodeSchema;
  ui?: NodeUI;
}