# dashboard/tmp/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Scheduler Public API (v1)
  version: 1.3.0
servers:
  - url: 'https://scheduler.example.com'
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workflows:
    get:
      tags:
        - Workflows
      summary: List stored workflows (paginated)
      operationId: listWorkflows
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowList'
    post:
      tags:
        - Workflows
      summary: Persist a workflow for editor storage (no versioning)
      operationId: persistWorkflow
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRef'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  '/api/v1/workflows/{workflowId}':
    get:
      tags:
        - Workflows
      summary: Read stored workflow (latest)
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Workflows
      summary: Soft delete workflow
      description: Marks the workflow record as deleted so it is hidden from listings and future reads.
      operationId: deleteWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '204':
          description: Workflow deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: 'Optional idempotency key for safe retries; if reused with a different body, return 409'
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflict (e.g., idempotency-key reuse with different body)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Authenticated but lacks required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Workflow:
      type: object
      additionalProperties: false
      required:
        - id
        - schemaVersion
        - metadata
        - nodes
        - edges
      properties:
        id:
          type: string
          format: uuid
          description: Workflow UUID
        schemaVersion:
          type: string
          description: e.g. "2025-10"
          pattern: '^[0-9]{4}-[0-9]{2}$'
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
        nodes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/WorkflowNode'
        edges:
          type: array
          minItems: 0
          items:
            $ref: '#/components/schemas/WorkflowEdge'
        tags:
          type: array
          description: Workflow-level tags.
          items:
            type: string
        previewImage:
          type: string
          description: Base64-encoded preview of the workflow canvas.
          nullable: true
        subgraphs:
          type: array
          description: Reusable workflow fragments (localized snapshots) that container nodes can reference.
          items:
            $ref: '#/components/schemas/WorkflowSubgraph'
    WorkflowRef:
      type: object
      required:
        - workflowId
      properties:
        workflowId:
          type: string
          format: uuid
    WorkflowSubgraph:
      type: object
      additionalProperties: false
      required:
        - id
        - definition
      properties:
        id:
          type: string
          description: Stable identifier referenced by container nodes.
        definition:
          $ref: '#/components/schemas/Workflow'
          description: Localized workflow snapshot backing this subgraph.
        metadata:
          $ref: '#/components/schemas/WorkflowSubgraphMetadata'
    WorkflowSubgraphMetadata:
      type: object
      additionalProperties: false
      properties:
        label:
          type: string
          description: Optional human-friendly label for the subgraph.
        description:
          type: string
        referenceWorkflowId:
          type: string
          format: uuid
          description: Original workflow id when this subgraph was created from a reference.
        referenceWorkflowName:
          type: string
        ownerId:
          type: string
        notes:
          type: string
...
    WorkflowList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Workflow'
        nextCursor:
          type: string
          nullable: true
...
    WorkflowMetadata:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        environment:
          type: string
        namespace:
          type: string
          description: Logical workflow namespace used for indexing/cross-workflow bindings. Defaults to "default".
        originId:
          type: string
          format: uuid
          description: Identifier linking versions of the same workflow. By default equals the workflow id.
        ownerId:
          type: string
          description: User id that owns the workflow definition.
        ownerName:
          type: string
          description: Display name of the owner if available.
          nullable: true
        createdBy:
          type: string
          description: User id that created the workflow definition.
        updatedBy:
          type: string
          description: User id that most recently updated the workflow definition.
      required:
        - name
...
    WorkflowNode:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - package
        - status
        - category
        - label
        - position
      properties:
        id:
          type: string
          format: uuid
          description: Node UUID
        type:
          type: string
          description: e.g. "playwright.open_page"
        role:
          type: string
          description: Execution role of the node.
          enum:
            - node
            - container
            - middleware
        package:
          $ref: '#/components/schemas/NodePackage'
        status:
          type: string
          description: Node lifecycle state.
          enum:
            - draft
            - published
            - deprecated
        category:
          type: string
          description: Group/category shown in the builder palette.
        label:
          type: string
        description:
          type: string
          description: Longer description of the node behaviour.
        tags:
          type: array
          description: Keywords for search/filter.
          items:
            type: string
        position:
          type: object
          required:
            - x
            - 'y'
          additionalProperties: false
          properties:
            x:
              type: number
            'y':
              type: number
        parameters:
          type: object
          description: Default parameter payload seeded from the manifest schema.
          additionalProperties: true
        results:
          type: object
          description: Default results payload seeded from the manifest schema.
          additionalProperties: true
        state:
          $ref: '#/components/schemas/WorkflowNodeState'
          description: Platform-managed execution hints surfaced in the builder (optional).
        schema:
          type: object
          description: JSON Schemas for parameters/results.
          additionalProperties: false
          properties:
            parameters:
              type: object
              additionalProperties: true
            results:
              type: object
              additionalProperties: true
        ui:
          $ref: '#/components/schemas/NodeUI'
        middlewares:
          type: array
          description: Ordered list of middleware definitions attached to this node.
          items:
            $ref: '#/components/schemas/WorkflowMiddleware'
    WorkflowMiddleware:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - package
        - status
        - category
        - label
      properties:
        id:
          type: string
          format: uuid
          description: Middleware UUID
        type:
          type: string
          description: e.g. "middleware.loop"
        role:
          type: string
          description: Execution role of the middleware (always middleware).
          enum:
            - middleware
        package:
          $ref: '#/components/schemas/NodePackage'
        status:
          type: string
          description: Middleware lifecycle state.
          enum:
            - draft
            - published
            - deprecated
        category:
          type: string
          description: Group/category shown in the builder palette.
        label:
          type: string
        description:
          type: string
          description: Longer description of the middleware behaviour.
        tags:
          type: array
          description: Keywords for search/filter.
          items:
            type: string
        parameters:
          type: object
          description: Default parameter payload seeded from the manifest schema.
          additionalProperties: true
        results:
          type: object
          description: Default results payload seeded from the manifest schema.
          additionalProperties: true
        state:
          $ref: '#/components/schemas/WorkflowNodeState'
          description: Platform-managed execution hints surfaced in the builder (optional).
        schema:
          type: object
          description: JSON Schemas for parameters/results.
          additionalProperties: false
          properties:
            parameters:
              type: object
              additionalProperties: true
            results:
              type: object
              additionalProperties: true
        ui:
          $ref: '#/components/schemas/NodeUI'
    NodePackage:
      type: object
      additionalProperties: false
      required:
        - name
        - version
      properties:
        name:
          type: string
        version:
          type: string
    NodeUI:
      type: object
      additionalProperties: false
      properties:
        inputPorts:
          type: array
          items:
            $ref: '#/components/schemas/UIPort'
        outputPorts:
          type: array
          items:
            $ref: '#/components/schemas/UIPort'
        widgets:
          type: array
          items:
            $ref: '#/components/schemas/UIWidget'
...
    WorkflowNodeState:
      type: object
      description: Scheduler-owned runtime hints attached to a workflow node during builder playback or run overlays.
      additionalProperties: true
      properties:
        stage:
          type: string
          description: 'Current scheduler-reported stage (e.g. idle, queued, running, completed, failed).'
        progress:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
          description: Optional 0-1 normalized progress indicator.
        lastUpdatedAt:
          type: string
          format: date-time
          nullable: true
          description: ISO-8601 timestamp for the last state transition.
        message:
          type: string
          nullable: true
          description: Human-friendly hint surfaced to the UI.
        error:
          $ref: '#/components/schemas/ResultError'
          nullable: true
          description: Optional terminal error surfaced when the node fails.
    UIPort:
      type: object
      additionalProperties: false
      required:
        - key
        - label
        - binding
      properties:
        key:
          type: string
        label:
          type: string
        binding:
          $ref: '#/components/schemas/UIBinding'
    UIBinding:
      type: object
      additionalProperties: false
      required:
        - path
        - mode
      properties:
        path:
          type: string
          description: e.g. parameters.url / results.status
        mode:
          type: string
          enum:
            - read
            - write
            - two_way
        prefix:
          type: string
          description: Optional textual prefix (e.g. '@welcomeJourney.stage2.#notifyCustomer') that scopes the binding before the JSON pointer root.
        scope:
          $ref: '#/components/schemas/BindingScope'
    BindingScope:
      type: object
      additionalProperties: false
      description: Canonical representation of a binding prefix used to resolve nodes across workflows or subgraphs.
      properties:
        kind:
          type: string
          enum:
            - local
            - subgraph
          description: 'local = default workflow, subgraph = inline subgraph alias chain.'
        subgraphAliases:
          type: array
          description: Ordered list of inline subgraph aliases traversed to reach the node.
          items:
            type: string
        nodeId:
          type: string
          format: uuid
          description: Explicit node identifier within the resolved scope (used for '#nodeId' prefixes or workflow-level references).
        prefix:
          type: string
          description: Raw prefix string captured before parsing (mirrors UIBinding.prefix for auditing/backwards compatibility).
    UIWidget:
      type: object
      additionalProperties: false
      required:
        - key
        - label
        - component
        - binding
      properties:
        key:
          type: string
        label:
          type: string
        component:
          type: string
          description: string|number|enum|json...
        binding:
          $ref: '#/components/schemas/UIBinding'
        options:
          type: object
          additionalProperties: true
    WorkflowEdge:
      type: object
      additionalProperties: false
      required:
        - id
        - source
        - target
      properties:
        id:
          type: string
        source:
          $ref: '#/components/schemas/EdgeEndpoint'
        target:
          $ref: '#/components/schemas/EdgeEndpoint'
    EdgeEndpoint:
      type: object
      additionalProperties: false
      required:
        - node
        - port
      properties:
        node:
          type: string
          format: uuid
          description: Node UUID
        port:
          type: string
          description: Port key defined by the node UI
...
    ResultError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
...